Restructure location plugin build
Restructured to better fit existing build structure
