+++ qgroundcontrol.pro
@@ -19,8 +19,14 @@
 
 TEMPLATE =  subdirs
 CONFIG  +=  ordered
-SUBDIRS  =  libs/QtLocationQGC
-SUBDIRS +=  ./qgcsystem.pro
+SUBDIRS  =  ./QGCLocationPlugin.pro
+SUBDIRS +=  ./QGCApplication.pro
 
-qgcsystem.depends = QtLocationQGC
+QGCApplication.depends = QGCLocationPlugin
+
+message(Qt version $$[QT_VERSION])
+
+!equals(QT_MAJOR_VERSION, 5) | !greaterThan(QT_MINOR_VERSION, 3) {
+    error("Unsupported Qt version, 5.4+ is required")
+}
