+++ src/QtLocationPlugin/OpenPilotMaps.cc
