+++ src/QtLocationPlugin/qgc_maps_plugin.json
