+++ src/QtLocationPlugin/qgeocodingmanagerengineqgc.h
