+++ src/QtLocationPlugin/qgeomapreplyqgc.cpp
